#142  [Timeframe manager] Implementation
